+++ uniform/src/main/java/org/comroid/restless/REST.java
@@ -3,7 +3,7 @@ package org.comroid.restless;
 import com.google.common.flogger.FluentLogger;
 import org.comroid.common.Polyfill;
 import org.comroid.common.func.Invocable;
-import org.comroid.common.iter.Span;
+import org.comroid.common.iter.span.Span;
 import org.comroid.restless.endpoint.CompleteEndpoint;
 import org.comroid.restless.endpoint.RatelimitedEndpoint;
 import org.comroid.restless.endpoint.RestEndpoint;
