+++ uniform/src/main/java/org/comroid/restless/server/Ratelimiter.java
@@ -1,7 +1,7 @@
 package org.comroid.restless.server;
 
 import com.google.common.flogger.FluentLogger;
-import org.comroid.common.iter.Span;
+import org.comroid.common.iter.span.Span;
 import org.comroid.restless.REST;
 import org.comroid.restless.endpoint.RatelimitedEndpoint;
