+++ uniform/src/main/java/org/comroid/varbind/model/AbstractObjectBind.java
@@ -3,7 +3,8 @@ package org.comroid.varbind.model;
 import java.util.function.BiFunction;
 
 import org.comroid.common.Polyfill;
-import org.comroid.common.iter.Span;
+import org.comroid.common.iter.span.BasicSpan;
+import org.comroid.common.iter.span.Span;
 import org.comroid.uniform.node.UniObjectNode;
 import org.comroid.varbind.bind.GroupBind;
 import org.comroid.varbind.bind.VarBind;
@@ -29,12 +30,12 @@ public abstract class AbstractObjectBind<EXTR, DPND, REMAP> implements VarBind<E
     }
 
     @Override
-    public final Span<EXTR> extract(UniObjectNode node) {
+    public final BasicSpan<EXTR> extract(UniObjectNode node) {
         return extractor.apply(node, fieldName);
     }
 
     @Override
-    public final REMAP finish(Span<REMAP> parts) {
+    public final REMAP finish(BasicSpan<REMAP> parts) {
         return parts.get();
     }
