+++ util/src/main/java/org/comroid/common/iter/span/ListSpan.java
@@ -1,0 +1,6 @@
+package org.comroid.common.iter.span;
+
+import java.util.List;
+
+public interface ListSpan<T> extends Span<T>, List<T> {
+}
