+++ util/src/main/java/org/comroid/common/iter/span/SetSpan.java
@@ -1,0 +1,6 @@
+package org.comroid.common.iter.span;
+
+import java.util.Set;
+
+public interface SetSpan<T> extends Span<T>, Set<T> {
+}
